Rename ResolveFieldContextAs to ResolveFieldContextAdapter
